Update backprop.py
After producing visualizations, these hooks will remain registered with the model, and when a user tries to backprop any parameter with gradient disabled, it will cause problems. Specifically, `grad_in[0]` in `_record_gradients` will be None, and `_record_output` may not be triggered. We can remind users to unregister those hooks after using flashtorch, but a simple thing that we could do is to skip those hooks by based on some checks.